DEV: Create ReplaceSysIoContextManager
Extracts the action of replacing sys.std* handles into a
separate context manager, so that alternative implementations
can re-use this component with other streams, or create wrappers
that do not use this functionality.
